+++ iodine-bukkit/src/main/java/hu/trigary/iodine/bukkit/gui/element/ButtonGuiElementImpl.java
@@ -29,7 +29,7 @@ public class ButtonGuiElementImpl extends GuiElementImpl<ButtonGuiElement> imple
 	 * @param internalId the internal ID of this element
 	 * @param id the API-friendly ID of this element
 	 */
-	public ButtonGuiElementImpl(@NotNull IodineGuiImpl gui, int internalId, @NotNull Object id) {
+	public ButtonGuiElementImpl(@NotNull IodineGuiImpl gui, short internalId, @NotNull Object id) {
 		super(gui, GuiElementType.BUTTON, internalId, id);
 	}
