+++ iodine-bukkit/src/main/java/hu/trigary/iodine/bukkit/gui/element/SliderGuiElementImpl.java
@@ -33,7 +33,7 @@ public class SliderGuiElementImpl extends GuiElementImpl<SliderGuiElement> imple
 	 * @param internalId the internal ID of this element
 	 * @param id the API-friendly ID of this element
 	 */
-	public SliderGuiElementImpl(@NotNull IodineGuiImpl gui, int internalId, @NotNull Object id) {
+	public SliderGuiElementImpl(@NotNull IodineGuiImpl gui, short internalId, @NotNull Object id) {
 		super(gui, GuiElementType.SLIDER, internalId, id);
 	}
